% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.fcc.num.packages-package.R
\docType{package}
\name{dc.fcc.num.packages-package}
\alias{dc.fcc.num.packages}
\alias{dc.fcc.num.packages-package}
\title{dc.fcc.num.packages: Social Data Commons: Digital Communcations: Number of Broadband Providers}
\description{
Allows user to easily get number of broadband provider data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Zachary Alerte \email{rsu8zv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
